+++ Engine/Paddle.h
@@ -8,16 +8,16 @@ public:
 	void MoveDown();
 	void Update();
 	static constexpr int width = 10;
-	int GetY(){ return y; }
+	int GetY(){ return static_cast<int>(y); }
 	int GetX(){ return x; }
 	int GetMovement(){ return movement; }
-	int GetHeight(){ return height; }
+	float GetHeight(){ return height; }
 	void Reset();
 	void Shrink();
 private:
 	const int x;
-	int y;
-	int yOld;
+	float y;
+	float yOld;
 	int movement;
-	int height;
+	float height;
 };
